+++ migrations/2025-07-17-161600_add-stadium-and-cheers/down.sql
@@ -1,4 +1,4 @@
 -- This file should undo anything in `up.sql`
 
-alter table data.games drop column stadium;
+alter table data.games drop column stadium_name;
 alter table data.events drop cheer;