+++ migrations/2025-07-17-161600_add-stadium-and-cheers/up.sql
@@ -1,4 +1,4 @@
 -- Your SQL goes here
 
-alter table data.games add column stadium text; -- null = pre-stadiums game
+alter table data.games add column stadium_name text; -- null = pre-stadiums game
 alter table data.events add cheer text; -- null = no cheer